Saving to a metadata file.
For now we are providing a saving capability, no exception handling yet, not saved or unsaved state management. The file is being saved next to where hardocs.json lives, and this path is hardcoded. In the future this will change, and allow for different configurations.

hardocs.json will be config.json per projects, and users can change and configure this file, including specifying which standard is being used, what is the name of the metadata file etc.

NEXT STEPS
- handle saved unsaved
- Load into the editor the real metadata and handle the situation where there might be no metadata.