Added the directory of the standard and also the standard. Having the standard in the state might come handy for various reasons, the most obvious to identify which kind of standard the metadata is using, perhaps the name or the reference should enough instead of saving the entire standard.
Issue:
Didnt manage to have the persisted or cached working directory for the standard. Once I open a new folder, and then go back to select a standard the reference path taken to open a file is from the previous project. I need to dig into how this is done